# client/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "GPG Signing Service API"
  },
  "components": {
    "schemas": {
...
      "PublicKeyResponse": {
        "type": "string",
        "example": "-----BEGIN PGP PUBLIC KEY BLOCK-----\n..."
      },
      "ErrorCode": {
        "type": "string",
        "enum": [
          "AUTH_MISSING",
          "AUTH_INVALID",
          "KEY_NOT_FOUND",
          "KEY_PROCESSING_ERROR",
          "KEY_LIST_ERROR",
          "KEY_UPLOAD_ERROR",
          "KEY_DELETE_ERROR",
          "SIGN_ERROR",
          "RATE_LIMIT_ERROR",
          "RATE_LIMITED",
          "INVALID_REQUEST",
          "AUDIT_ERROR",
          "NOT_FOUND",
          "INTERNAL_ERROR"
        ]
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "minLength": 1
          },
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "requestId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "error",
          "code"
        ]
      },
      "SignResponse": {
        "type": "string",
        "example": "-----BEGIN PGP SIGNATURE-----\n..."
      },
      "RateLimitError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "minLength": 1
          },
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "retryAfter": {
            "type": "integer",
            "minimum": 0,
            "exclusiveMinimum": true
          }
        },
        "required": [
          "error",
          "code",
          "retryAfter"
        ]
      },
      "SignRequest": {
        "type": "string",
        "minLength": 1,
        "example": "tree 29ff16c9c14e2652b22f8b78bb08a5a07930c147\nparent ..."
      },
...
    },
...
    "securitySchemes": {
      "oidcAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "OIDC token from GitHub Actions or GitLab CI"
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Admin token for /admin/* endpoints"
      }
    }
  },
  "paths": {
...
    "/public-key": {
      "get": {
        "summary": "Get public key",
        "description": "Get the public key for signature verification",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "example": "A1B2C3D4E5F6G7H8"
            },
            "required": false,
            "name": "keyId",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Public Key",
            "content": {
              "application/pgp-keys": {
                "schema": {
                  "$ref": "#/components/schemas/PublicKeyResponse"
                }
              }
            }
          },
          "404": {
            "description": "Key not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/sign": {
      "post": {
        "summary": "Sign commit data",
        "description": "Sign git commit data using the stored GPG key",
        "security": [
          {
            "oidcAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "example": "A1B2C3D4E5F6G7H8"
            },
            "required": false,
            "name": "keyId",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": false,
            "name": "X-Request-ID",
            "in": "header"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "text/plain": {
              "schema": {
                "$ref": "#/components/schemas/SignRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "PGP Signature",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SignResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Key not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RateLimitError"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}